+++ run_functional_tests
@@ -1,10 +1,9 @@
 #!/bin/bash
 #
-# Simple script to run all tests with multiple test servers
-# across all supported Python versions
+# Simple script to run all functional tests with multiple test servers
 #
 
-# Default test server running latest self-hosted site
+# Test server running latest self-hosted site
 
 tput setaf 3
 echo
@@ -12,7 +11,7 @@ echo "Testing latest self-hosted site..."
 tput sgr0
 export OPENPHOTO_TEST_CONFIG=test
 unset OPENPHOTO_TEST_SERVER_API
-tox $@
+python -m unittest discover --catch tests/functional
 
 # Test server running APIv1 OpenPhoto instance
 tput setaf 3
@@ -21,7 +20,7 @@ echo "Testing APIv1 self-hosted site..."
 tput sgr0
 export OPENPHOTO_TEST_CONFIG=test-apiv1
 export OPENPHOTO_TEST_SERVER_API=1
-tox $@
+python -m unittest discover --catch tests/functional
 
 # Test account on hosted trovebox.com site
 tput setaf 3
@@ -30,5 +29,5 @@ echo "Testing latest hosted site..."
 tput sgr0
 export OPENPHOTO_TEST_CONFIG=test-hosted
 unset OPENPHOTO_TEST_SERVER_API
-tox $@
+python -m unittest discover --catch tests/functional
