+++ tests/functional/api_versions/test_v1.py
@@ -1,4 +1,4 @@
-from tests import test_albums, test_photos, test_tags
+from tests.functional import test_albums, test_photos, test_tags
 
 class TestAlbumsV1(test_albums.TestAlbums):
     api_version = 1
