+++ tests/functional/test_albums.py
@@ -1,6 +1,6 @@
-import tests.test_base
+from tests.functional import test_base
 
-class TestAlbums(tests.test_base.TestBase):
+class TestAlbums(test_base.TestBase):
     testcase_name = "album API"
 
     def test_create_delete(self):
