+++ tox.ini
@@ -2,10 +2,15 @@
 envlist = py26, py27, py33
 
 [testenv]
-commands = python -m unittest discover --catch
+commands = python -m unittest discover --catch tests/unit
+deps =
+    mock
+    httpretty
 
 [testenv:py26]
-commands = unit2 discover --catch
+commands = unit2 discover --catch tests/unit
 deps =
+    mock
+    httpretty
     unittest2
     discover
